+++ Sources/SwifterSwift/MapKit/MKMapViewExtensions.swift
@@ -45,7 +45,26 @@ public extension MKMapView {
         return annotationView
     }
 
+    /// SwifterSwift: Zooms in on multiple mapView coordinates.
+    ///
+    /// - Parameters:
+    ///   - coordinates: Gets the array of type CLLocationCoordinate2D.
+    ///   - meter: If arrays have a single item, they take the value of meters (Double). The map zooms in at the given meters.
+    ///   - edgePadding: The amount of additional space (measured in screen points) to make visible around the specified rectangle
+    ///   - animated: The animation control takes the Boolean value. Enter the true value for zooming with the animation.
+    func zoom(to coordinates: [CLLocationCoordinate2D], meter: Double, edgePadding: EdgeInsets, animated: Bool) {
+        guard !coordinates.isEmpty else { return }
+
+        if coordinates.count == 1 {
+            let coordinateRegion = MKCoordinateRegion(center: coordinates.first!, latitudinalMeters: meter, longitudinalMeters: meter)
+            setRegion(coordinateRegion, animated: true)
+        } else {
+            let mkPolygon = MKPolygon(coordinates: coordinates, count: coordinates.count)
+            setVisibleMapRect(mkPolygon.boundingMapRect, edgePadding: edgePadding, animated: animated)
+        }
+    }
 }
+
 #endif
 
 #endif